fixed fontconvert for MacOS homebrew and the GFXglyph structure to correctly handle large fonts

+++ gfxfont.h
@@ -10,11 +10,11 @@
 /// Font data stored PER GLYPH
 typedef struct {
   uint16_t bitmapOffset; ///< Pointer into GFXfont->bitmap
-  uint8_t width;         ///< Bitmap dimensions in pixels
-  uint8_t height;        ///< Bitmap dimensions in pixels
-  uint8_t xAdvance;      ///< Distance to advance cursor (x axis)
-  int8_t xOffset;        ///< X dist from cursor pos to UL corner
-  int8_t yOffset;        ///< Y dist from cursor pos to UL corner
+  uint16_t width;         ///< Bitmap dimensions in pixels
+  uint16_t height;        ///< Bitmap dimensions in pixels
+  uint16_t xAdvance;      ///< Distance to advance cursor (x axis)
+  int16_t xOffset;        ///< X dist from cursor pos to UL corner
+  int16_t yOffset;        ///< Y dist from cursor pos to UL corner
 } GFXglyph;
 
 /// Data stored for FONT AS A WHOLE
@@ -23,7 +23,7 @@ typedef struct {
   GFXglyph *glyph;  ///< Glyph array
   uint16_t first;   ///< ASCII extents (first char)
   uint16_t last;    ///< ASCII extents (last char)
-  uint8_t yAdvance; ///< Newline distance (y axis)
+  uint16_t yAdvance; ///< Newline distance (y axis)
 } GFXfont;
 
 #endif // _GFXFONT_H_
